requirements.txt
annotated-types==0.7.0
anyio==4.7.0
beanie==1.28.0
click==8.1.7
dnspython==2.7.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
lazy-model==0.2.0
motor==3.6.0
passlib==1.7.4
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pymongo==4.9.2
sniffio==1.3.1
starlette==0.41.3
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.32.1
